+++ lib/apathy_drive/commands/move.ex
@@ -107,7 +107,7 @@ defmodule ApathyDrive.Commands.Move do
   end
 
   def execute(%Room{} = room, %{} = character, %{"kind" => "Alignment"} = room_exit, reattempt) do
-    if character.alignment in room_exit["allowed"] do
+    if Character.alignment(character) in room_exit["allowed"] do
       execute(room, room.mobiles[character.ref], Map.put(room_exit, "kind", "Normal"), reattempt)
     else
       Mobile.send_scroll(character, "<p>A strange power holds you back!</p>")