+++ lib/apathy_drive/mobile.ex
@@ -17,6 +17,7 @@ defprotocol ApathyDrive.Mobile do
   def dodge_at_level(mobile, level, room)
   def enough_mana_for_ability?(mobile, ability)
   def enter_message(mobile)
+  def evil_points(mobile, attacker)
   def exhausted(mobile)
   def exit_message(mobile)
   def has_ability?(mobile, ability_name)