+++ priv/repo/migrations/20190824170834_character_alignment_tweaks.exs
@@ -1,0 +1,12 @@
+defmodule ApathyDrive.Repo.Migrations.CharacterAlignmentTweaks do
+  use Ecto.Migration
+
+  def change do
+    alter table(:characters) do
+      remove(:alignment)
+      remove(:bounty)
+      add(:evil_points, :float, default: 0)
+      add(:last_evil_action_at, :utc_datetime)
+    end
+  end
+end